Rename dbserver.js to server.js
